TST: create_2d_test_case.py
